Minor Changes to utils.
