+++ _CORE_PROJECTS/HaleyHelpers/HaleyHelpers.csproj
@@ -10,19 +10,19 @@
     <Authors>Lingam</Authors>
     <Copyright>Haley 2020</Copyright>
     <PackageIcon>Haley.png</PackageIcon>
-    <AssemblyVersion>1.0.10.0</AssemblyVersion>
-    <FileVersion>1.0.10.0</FileVersion>
+    <AssemblyVersion>1.0.11.0</AssemblyVersion>
+    <FileVersion>1.0.11.0</FileVersion>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <Product>Haley.Helpers</Product>
     <PackageId>Haley.Helpers</PackageId>
-    <Version>1.0.10</Version>
-    <AssemblyName>Haley.Helpers.Standard</AssemblyName>
+    <Version>1.0.11</Version>
+    <AssemblyName>Haley.Helpers</AssemblyName>
     <PackageProjectUrl>https://github.com/rmsmech/HaleyProject</PackageProjectUrl>
     <RepositoryUrl>https://github.com/rmsmech/HaleyProject</RepositoryUrl>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
 	  <IncludeSymbols>true</IncludeSymbols>
 	  <SymbolPackageFormat>snupkg</SymbolPackageFormat>
-	  <PackageReleaseNotes>Modifications in hashUtils</PackageReleaseNotes>
+	  <PackageReleaseNotes>Added RandomBool generator to generalutils.</PackageReleaseNotes>
   </PropertyGroup>
 
   <ItemGroup>
